improved memory management
- migliorata la gestione dello spazio di memoria durante l'inizializzazione del file system e durante la creazione di un file
- modificati alcuni comandi (create e delete diventano mk e rm)
- modificata struttura dati FileEntry (short diventano int -> allineamento dei file più efficiente nella FAT)

+++ fs_struct.h
@@ -19,8 +19,8 @@ typedef struct {
     char name[16];
     int start_block;
     int size;
-    short is_used;
-    short is_directory;
+    int is_used;
+    int is_directory;
 } FileEntry;
 
 typedef struct {
